+++ src/save.cpp
@@ -899,11 +899,11 @@ bool get_char(bool *generate) {
                 if (c_ptr >= &cave[MAX_HEIGHT][0]) {
                     goto error;
                 }
-                c_ptr->fval = char_tmp & 0xF;
-                c_ptr->lr = (char_tmp >> 4) & 0x1;
-                c_ptr->fm = (char_tmp >> 5) & 0x1;
-                c_ptr->pl = (char_tmp >> 6) & 0x1;
-                c_ptr->tl = (char_tmp >> 7) & 0x1;
+                c_ptr->fval = (uint8_t)(char_tmp & 0xF);
+                c_ptr->lr = (unsigned int)((char_tmp >> 4) & 0x1);
+                c_ptr->fm = (unsigned int)(char_tmp >> 5) & 0x1;
+                c_ptr->pl = (unsigned int)(char_tmp >> 6) & 0x1;
+                c_ptr->tl = (unsigned int)(char_tmp >> 7) & 0x1;
                 c_ptr++;
             }
             total_count += count;
